add some late-stage inlining mechanism
Because of our Sally shenanigans, and due to how What4 works, the
generated Sally terms contain many let-bindings for values that are
essentially constants (or fields of constants).

They make the output harder to read than it ought to.  This commit
implements an inefficient way of inlining those spurious let-bindings.
Ideally, it should be configurable via an option, since it could be
quite slow on very large files.