language-sally-0.4.0.0
This version introduces the `Language.Sally.TransitionSystem` module,
which used to be part of the `what4-transition-system` project but made
maintenance awful by introducing cycles in the dependencies.

The module could arguably be moved to its own Haskell package, but
currently it is not worth the effort to keep track of it separately.